[epociask/issue-34-hardcoded-invariant-support] Cleaned up code, added comments
