+++ internal/core/constants.go
@@ -27,54 +27,6 @@ func (n Network) String() string {
 	return UnknownType
 }
 
-// ComponentType ...
-type ComponentType uint8
-
-const (
-	Oracle ComponentType = iota + 1
-	Pipe
-	Aggregator
-)
-
-func (ct ComponentType) String() string {
-	switch ct {
-	case Oracle:
-		return "oracle"
-
-	case Pipe:
-		return "pipe"
-
-	case Aggregator:
-		return "aggregator"
-	}
-
-	return UnknownType
-}
-
-// PipelineType ...
-type PipelineType uint8
-
-const (
-	Backtest PipelineType = iota + 1
-	Live
-	MockTest
-)
-
-func (pt PipelineType) String() string {
-	switch pt {
-	case Backtest:
-		return "backtest"
-
-	case Live:
-		return "live"
-
-	case MockTest:
-		return "mocktest"
-	}
-
-	return UnknownType
-}
-
 type FetchType int
 
 const (
@@ -108,6 +60,7 @@ func (it InvariantType) String() string {
 	}
 }
 
+// ID keys used for logging
 const (
 	CUUIDKey logging.LogKey = "cuuid"
 	PUUIDKey logging.LogKey = "puuid"
