+++ internal/core/id.go
@@ -6,16 +6,19 @@ import (
 	"github.com/google/uuid"
 )
 
+// UUID ... third-party wrapper struct
 type UUID struct {
 	uuid.UUID
 }
 
+// newUUID ... Constructor
 func newUUID() UUID {
 	return UUID{
 		uuid.New(),
 	}
 }
 
+// nilUUID ... Returns a zero'd out 16 byte array
 func nilUUID() UUID {
 	return UUID{[16]byte{0}}
 }
@@ -67,12 +70,6 @@ type InvSessionUUID struct {
 	PID  InvSessionPID
 	UUID UUID
 }
-
-/*
-	NOTE: Pipelines that require a backfill will cause inaccurate collisions
-	within the pipeline DAG.
-
-*/
 
 // NOTE - This is useful for error handling with functions that
 // also return a ComponentID
@@ -190,6 +187,7 @@ func (id PipelineUUID) String() string {
 	)
 }
 
+// String ... Returns string representation of an invariant sesion PID
 func (id InvSessionPID) String() string {
 	return fmt.Sprintf("%s:%s:%s",
 		Network(id[0]).String(),
@@ -198,6 +196,7 @@ func (id InvSessionPID) String() string {
 	)
 }
 
+// String ... Returns string reprsentation of an invariant session UUID
 func (id InvSessionUUID) String() string {
 	return fmt.Sprintf("%s::%s",
 		id.PID.String(), id.UUID.ShortString())
